+++ NEWS.md
@@ -1,6 +1,6 @@
 # SciViews 1.9.0
 
--   A `global_entrace=` argument is added to `SciViews::R` in ordre to use the {rlang} condition messages (by default, only for error messages).
+-   A `global_entrace=` argument is added to `SciViews::R` in order to use the {rlang} condition messages (by default, only for error messages).
 
 -   `SciViews::R` now tries loading all required packages (it does not stop any more at the first failure). Consequently, `Install()` is more useful by installing all required packages at once, instead of one by one.